% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.R
\name{isochrone}
\alias{isochrone}
\title{Get Polygons of Isochrones}
\usage{
isochrone(
  loc,
  breaks = seq(from = 0, to = 60, length.out = 7),
  exclude,
  res = 30,
  method = "osrm_pkg",
  k,
  osrm.server = getOption("osrm.server"),
  osrm.profile = getOption("osrm.profile")
)
}
\arguments{
\item{loc}{origin point. \code{loc} can be: \itemize{
  \item a vector of coordinates (longitude and latitude, WGS 84),
  \item a data.frame of longitudes and latitudes (WGS 84),
  \item a matrix of longitudes and latitudes (WGS 84),
  \item an sfc object of type POINT,
  \item an sf object of type POINT.
}
If \code{loc} is a data.frame, a matrix, an sfc object or an sf object then
only the first row or element is considered.}

\item{breaks}{a numeric vector of break values to define isochrone areas,
in minutes.}

\item{exclude}{pass an optional "exclude" request option to the OSRM API.}

\item{res}{number of points used to compute isochrones, one side of the square
grid, the total number of points will be res*res. Increase res to obtain more
detailed isochrones.}

\item{method}{select the iso creation method. It can be "osrm_pkg" or "smooth". 
"osrm_pkg" uses the same algorithm than the osrm package. "smooth" uses a 
moving window with a gaussian blur. "smooth" is usefull to remove small 
patches of area hard to reach.}

\item{k}{size (sigma) of the gaussian moving window.}

\item{osrm.server}{the base URL of the routing server.
getOption("osrm.server") by default.}

\item{osrm.profile}{the routing profile to use, e.g. "car", "bike" or "foot"
(when using the routing.openstreetmap.de test server).
getOption("osrm.profile") by default.}
}
\value{
The output of this function is an sf MULTIPOLYGON of isochrones.\cr
It contains 3 fields: \itemize{
  \item id, an identifier
  \item isomin, the minimum value of the isochrone polygon in minutes
  \item isomax, the maximum value of the isochrone polygon in minutes
}
If loc is a vector, a data.frame or a matrix the coordinate
reference system (CRS) of the output is EPSG:4326 (WGS84).\cr
If loc is an sfc or sf object, the output has the same CRS
as loc.\cr
}
\description{
This function computes areas that are reachable within a
given time span from a point and returns the reachable regions as
polygons. These areas of equal travel time are called isochrones.
}
\examples{
\dontrun{
library(sf)
# Get isochones with lon/lat coordinates
iso <- isochrone(loc = c(13.43, 52.47), breaks = seq(0, 12, 2))
# Map
plot(iso["isomax"], breaks = sort(unique(c(iso$isomin, iso$isomax))))
}
}
